Split out DynamicFieldset test to own file
